Refactor Thing bindings
More closely reflects Producer interface on W3 Scripting API. Each
device type now defines a produce method that searches for devices
it can expose and builds the TD and a Flask blueprint. As it is a
blueprint for each device a database is no longer needed as the url
and address are set at the time of blueprint creation.