Move controllable parameters to config file
Will be installed at /etc/wot-network.ini. Uses configparser module
to parse this file and pulls out information like the db location,
hostnames etc. This will also be used to store the secrets for
generating API tokens.

To cater for this the logic for binding producers had to be modified
slightly.