Begin implementation of grouper service
Will allow things to be grouped together and abstracted under a single
set of interactions.

Trying out a new structure for flask apis: The main logic is in a
service which does not depend on Flask. The flask routes merely call
methods in the service, and translate caught exceptions into HTTP
errors.